requirements.txt
langchain
langchain-groq
openai
streamlit
newsapi-python
python-dotenv
fpdf2